mina_zkml::graph
Module model

This module defines the core structures and functions for representing and manipulating neural network graphs. It includes the main Model structure, which contains the parsed graph and variable visibility settings, as well as various enums and structs for representing nodes, operations, and connections within the graph.

Structs

Model

Main model structure containing the parsed graph and variable visibility settings

ParsedNodes

Represents the parsed neural network graph structure

RunArgs

Arguments for running the model

SerializableNode

Serializable version of Node that excludes TypedOp

SerializableOutletId

Serializable version of OutletId

VarVisibility

Controls visibility of variables in the model

Enums

InputMapping

Defines how inputs are mapped in subgraphs

NodeType

Represents different types of nodes in the graph

OperationType

Enum representing different types of operations that can be performed in the graph.

OutputMapping

Defines how outputs are mapped in subgraphs

Visibility

Variable visibility levels

Type Aliases

GraphLoadResult

Type alias for the graph loading result

Outlet

Represents a node output connection as (node_index, output_slot)