mina_zkml::graph
Module utilities

Functions

create_add_axis_node

Creates an AddAxis node with the given parameters.

create_const_node

Creates a Const node with the given id, shape, and values.

create_conv_node

Creates a Conv node with the given parameters.

create_elementwise_node

Creates an ElementWiseOp node with the given parameters.

create_gather_node

Creates a Gather node with the given parameters.

create_input_node

Creates an Input node with the given id and shape.

create_max_pool_node

Creates a MaxPool node with the given parameters.

create_node

Creates a new node with the given parameters.

create_reduce_node

Creates a Reduce node with the given parameters.

create_softmax_node

Creates a Softmax node with the given parameters.

create_typedbin_node

Creates a TypedBinOp node with the given parameters.

get_value_from_attributes

Retrieves a value from the attributes HashMap.

handle_pool_spec

Handles the PoolSpec attributes and inserts them into the provided attributes HashMap.

node_output_shapes

Retrieves the output shapes of a given ONNX node.